Attempted to fix flight trajectory calculations
Made some minor updates based on info from OpenRocket (starting mass = 0.921 kg; ending mass = 0.834 kg; cross sectional area). Added burnout logic - mass doesn't change and thrust goes to zero after burnout. Changed lines 131/135 to write "trajectory_fun()" in terms of state variables and not output variables to see if it would help with the calculations.

I tried messing with the lift vector to see if something would change, and when set to zero we get the expected parabolic shape of the trajectory. Maybe something is wrong here, but I was unable to figure it out on my own last night. Output is not really better than when Daniel worked on it, but I plan on trying again later this evening.