Updated CL and CD - got more realistic results
Updating CL < CD fixes the issues we were having with the plots looking crazy (previously we had CL > CD). plot(z, h) actually looks like a trajectory now; however, it doesn't seem to go as high as the results from OpenRocket. So we're def on the right track, but it may need some more tweaking.

Still need the optimal launch angle for given wind speed and the parachute modeling.