Minor tweaks; removed L, tried new psi
Dr. Heister said we could set lift equal to zero, so I removed those from the calculations. I also tried calculating "psi" using the other formulas given in the "trajectory help" doc on Brightspace, but this doesn't help the breakdown of the code for vertical launches.

The code also breaks down if Vw = 0.